+++ e-library/src/Components/button/button.jsx
@@ -1,13 +1,11 @@
 import React from 'react';
 
-function Button() {
-  return (
-    <div className="App">
-      <div>
-        <button type="button">BORROW</button>
-      </div>
+const Button = () => (
+  <div className="App">
+    <div>
+      <button type="button">BORROW</button>
     </div>
-  );
-}
+  </div>
+);
 
 export default Button;
