+++ e-library/src/Components/forms/createbook.jsx
@@ -1,59 +1,23 @@
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
-import FormData from 'form-data';
-import axios from '../api/axios';
-import { userInfo } from '../homepage/body';
-import useCustomHooks from './useCustomhooks';
+import useBookInputField from '../hooks/useBookInputField';
 
-export const bookData = {};
-
-export const CreateBook = () => {
-  const [book, formValue, bookfile] = useCustomHooks({});
-
-  const redirect = useNavigate();
-
-  const submit = async (e) => {
-    e.preventDefault();
-    const form = new FormData();
-    form.append('title', book.title);
-    form.append('price', book.price);
-    form.append('author', book.author);
-    form.append('description', book.description);
-    form.append('bookFile', book.fileUrl);
-    try {
-      const response = await axios.post('/api/books', form, {
-        headers: {
-          'Content-Type': 'multipart/form-data',
-          authorization: userInfo.token,
-        },
-      });
-      window.alert('Book has been created successfully');
-      redirect('/homepage');
-      bookData.file = response.data.createbook.fileUrl;
-      bookData.id = response.data.createbook.id;
-    } catch (err) {
-      if (err) {
-        window.alert('Failed to create book. Please try again');
-        redirect('/homepage');
-        console.log(err);
-      }
-    }
-  };
+const CreateBook = () => {
+  const [formValue, bookfile, submit] = useBookInputField();
   return (
     <div>
       <form onSubmit={submit}>
         <h2>Create your book</h2>
         <div className="input-box">
-          <input type="text" value={book.title} name="title" placeholder="Title" onChange={formValue} />
+          <input type="text" name="title" placeholder="Title" onChange={formValue} />
         </div>
         <div className="input-box">
-          <textarea placeholder="Description" value={book.description} name="description" rows="4" columns="50" onChange={formValue} />
+          <textarea placeholder="Description" name="description" rows="4" columns="50" onChange={formValue} />
         </div>
         <div className="input-box">
-          <input type="text" placeholder="Author" value={book.author} name="author" onChange={formValue} />
+          <input type="text" placeholder="Author" name="author" onChange={formValue} />
         </div>
         <div className="input-box">
-          <input type="text" placeholder="Price" value={book.price} name="price" onChange={formValue} />
+          <input type="text" placeholder="Price" name="price" onChange={formValue} />
         </div>
         <div className="input-box">
           <input type="file" onChange={bookfile} required />
@@ -65,3 +29,5 @@ export const CreateBook = () => {
     </div>
   );
 };
+
+export default CreateBook;
